Move HashKey to a separate file
